+++ processor/src/main/java/react4j/processor/Constants.java
@@ -6,6 +6,7 @@ final class Constants
   static final String ERROR_INFO_CLASSNAME = "react4j.ReactErrorInfo";
   static final String KEYED_CLASSNAME = "react4j.Keyed";
   static final String PROP_ANNOTATION_CLASSNAME = "react4j.annotations.Prop";
+  static final String SUPPRESS_REACT4J_WARNINGS_ANNOTATION_CLASSNAME = "react4j.annotations.SuppressReact4jWarnings";
   static final String PROP_REF_ANNOTATION_CLASSNAME = "react4j.annotations.PropRef";
   static final String PROP_DEFAULT_ANNOTATION_CLASSNAME = "react4j.annotations.PropDefault";
   static final String PROP_VALIDATE_ANNOTATION_CLASSNAME = "react4j.annotations.PropValidate";